+++ bindings/nostr-ffi/src/event/raw.rs
@@ -82,7 +82,7 @@ impl RawEvent {
         self.inner.clone().into()
     }
 
-    pub fn as_json(&self) -> String {
-        self.inner.as_json()
+    pub fn as_json(&self) -> Result<String> {
+        Ok(self.inner.try_as_json()?)
     }
 }