+++ bindings/nostr-ffi/src/nips/nip15.rs
@@ -132,8 +132,8 @@ impl StallData {
         self.inner.clone().into()
     }
 
-    pub fn as_json(&self) -> String {
-        self.inner.as_json()
+    pub fn as_json(&self) -> Result<String> {
+        Ok(self.inner.try_as_json()?)
     }
 }