+++ bindings/nostr-ffi/src/types/filter.rs
@@ -392,8 +392,8 @@ impl Filter {
         })
     }
 
-    pub fn as_json(&self) -> String {
-        self.inner.as_json()
+    pub fn as_json(&self) -> Result<String> {
+        Ok(self.inner.try_as_json()?)
     }
 }