+++ bindings/nostr-ffi/src/types/metadata.rs
@@ -113,8 +113,8 @@ impl Metadata {
         self.inner.clone().into()
     }
 
-    pub fn as_json(&self) -> String {
-        self.inner.as_json()
+    pub fn as_json(&self) -> Result<String> {
+        Ok(self.inner.try_as_json()?)
     }
 
     pub fn set_name(self: Arc<Self>, name: String) -> Self {