+++ bindings/nostr-js/src/event/mod.rs
@@ -102,7 +102,7 @@ impl JsEvent {
     }
 
     #[wasm_bindgen(js_name = asJson)]
-    pub fn as_json(&self) -> String {
-        self.inner.as_json()
+    pub fn as_json(&self) -> Result<String> {
+        self.inner.try_as_json().map_err(into_err)
     }
 }