+++ bindings/nostr-js/src/message/relay.rs
@@ -83,7 +83,7 @@ impl JsRelayMessage {
     }
 
     #[wasm_bindgen(js_name = asJson)]
-    pub fn as_json(&self) -> String {
-        self.inner.as_json()
+    pub fn as_json(&self) -> Result<String> {
+        self.inner.try_as_json().map_err(into_err)
     }
 }